requirements.txt
transformers==4.35.0
torch==2.1.0
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.0
datasets==2.14.0
PyPDF2==3.0.1
python-docx==1.1.0